Decoupling shared utility functions from test files
