+++ tests/e2e/staging_env_basic.go
@@ -41,7 +41,6 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework/manifest"
 	fnodes "k8s.io/kubernetes/test/e2e/framework/node"
 	fpod "k8s.io/kubernetes/test/e2e/framework/pod"
-	e2eoutput "k8s.io/kubernetes/test/e2e/framework/pod/output"
 	fpv "k8s.io/kubernetes/test/e2e/framework/pv"
 	fss "k8s.io/kubernetes/test/e2e/framework/statefulset"
 
@@ -1257,115 +1256,3 @@ var _ = ginkgo.Describe("[csi-supervisor-staging] Tests for WCP env with minimal
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	})
 })
-
-// createPODandVerifyVolumeMount this method creates Pod and verifies VolumeMount
-func createPODandVerifyVolumeMountWithoutF(ctx context.Context, client clientset.Interface,
-	namespace string, pvclaim *v1.PersistentVolumeClaim, volHandle string) (*v1.Pod, string) {
-	// Create a Pod to use this PVC, and verify volume has been attached
-	ginkgo.By("Creating pod to attach PV to the node")
-	pod, err := createPod(ctx, client, namespace, nil, []*v1.PersistentVolumeClaim{pvclaim}, false, execCommand)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-	defer func() {
-		err = client.CoreV1().Pods(namespace).Delete(ctx, pod.Name, *metav1.NewDeleteOptions(0))
-		if !apierrors.IsNotFound(err) {
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		}
-	}()
-
-	var exists bool
-	var vmUUID string
-	ginkgo.By(fmt.Sprintf("Verify volume: %s is attached to the node: %s", volHandle, pod.Spec.NodeName))
-
-	annotations := pod.Annotations
-	vmUUID, exists = annotations[vmUUIDLabel]
-	gomega.Expect(exists).To(gomega.BeTrue(), fmt.Sprintf("Pod doesn't have %s annotation", vmUUIDLabel))
-
-	framework.Logf("VMUUID : %s", vmUUID)
-	isDiskAttached, err := e2eVSphere.isVolumeAttachedToVM(client, volHandle, vmUUID)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	gomega.Expect(isDiskAttached).To(gomega.BeTrue(),
-		"Volume is not attached to the node volHandle: %s, vmUUID: %s", volHandle, vmUUID)
-
-	ginkgo.By("Verify the volume is accessible and filesystem type is as expected")
-	_, err = e2eoutput.LookForStringInPodExec(namespace, pod.Name,
-		[]string{"/bin/cat", "/mnt/volume1/fstype"}, "", time.Minute)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-	return pod, vmUUID
-}
-
-// getFSSizeMb returns filesystem size in Mb
-func getFSSizeMbWithoutF(namespace string, pod *v1.Pod) (int64, error) {
-	var output string
-	var err error
-
-	cmd := []string{"exec", pod.Name, "--namespace=" + namespace, "--", "/bin/sh", "-c", "df -Tkm | grep /mnt/volume1"}
-	output = e2ekubectl.RunKubectlOrDie(namespace, cmd...)
-	gomega.Expect(strings.Contains(output, ext4FSType)).NotTo(gomega.BeFalse())
-
-	arrMountOut := strings.Fields(string(output))
-	if len(arrMountOut) <= 0 {
-		return -1, fmt.Errorf("error when parsing output of `df -T`. output: %s", string(output))
-	}
-	var devicePath, strSize string
-	devicePath = arrMountOut[0]
-	if devicePath == "" {
-		return -1, fmt.Errorf("error when parsing output of `df -T` to find out devicePath of /mnt/volume1. output: %s",
-			string(output))
-	}
-	strSize = arrMountOut[2]
-	if strSize == "" {
-		return -1, fmt.Errorf("error when parsing output of `df -T` to find out size of /mnt/volume1: output: %s",
-			string(output))
-	}
-
-	intSizeInMb, err := strconv.ParseInt(strSize, 10, 64)
-	if err != nil {
-		return -1, fmt.Errorf("failed to parse size %s into int size", strSize)
-	}
-
-	return intSizeInMb, nil
-}
-
-// increaseSizeOfPvcAttachedToPod this method increases the PVC size, which is attached to POD
-func increaseSizeOfPvcAttachedToPodWithoutF(client clientset.Interface,
-	namespace string, pvclaim *v1.PersistentVolumeClaim, pod *v1.Pod) {
-	var originalSizeInMb int64
-	var err error
-	//Fetch original FileSystemSize
-	ginkgo.By("Verify filesystem size for mount point /mnt/volume1 before expansion")
-	originalSizeInMb, err = getFSSizeMbWithoutF(namespace, pod)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-	//resize PVC
-	// Modify PVC spec to trigger volume expansion
-	ginkgo.By("Expanding current pvc")
-	currentPvcSize := pvclaim.Spec.Resources.Requests[v1.ResourceStorage]
-	newSize := currentPvcSize.DeepCopy()
-	newSize.Add(resource.MustParse("1Gi"))
-	framework.Logf("currentPvcSize %v, newSize %v", currentPvcSize, newSize)
-	pvclaim, err = expandPVCSize(pvclaim, newSize, client)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	gomega.Expect(pvclaim).NotTo(gomega.BeNil())
-
-	ginkgo.By("Waiting for file system resize to finish")
-	pvclaim, err = waitForFSResize(pvclaim, client)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-	pvcConditions := pvclaim.Status.Conditions
-	expectEqual(len(pvcConditions), 0, "pvc should not have conditions")
-
-	var fsSize int64
-	ginkgo.By("Verify filesystem size for mount point /mnt/volume1")
-	fsSize, err = getFSSizeMbWithoutF(namespace, pod)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	framework.Logf("File system size after expansion : %v", fsSize)
-	// Filesystem size may be smaller than the size of the block volume
-	// so here we are checking if the new filesystem size is greater than
-	// the original volume size as the filesystem is formatted for the
-	// first time
-	gomega.Expect(fsSize).Should(gomega.BeNumerically(">", originalSizeInMb),
-		fmt.Sprintf("error updating filesystem size for %q. Resulting filesystem size is %d", pvclaim.Name, fsSize))
-	ginkgo.By("File system resize finished successfully")
-}
