+++ tests/e2e/volume_expansion_utils.go
@@ -1,0 +1,106 @@
+/*
+Copyright 2025 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package e2e
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/onsi/ginkgo/v2"
+	"github.com/onsi/gomega"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	clientset "k8s.io/client-go/kubernetes"
+	"k8s.io/kubernetes/test/e2e/framework"
+	e2ekubectl "k8s.io/kubernetes/test/e2e/framework/kubectl"
+)
+
+// getFSSizeMb returns filesystem size in Mb
+func getFSSizeMbWithoutF(namespace string, pod *v1.Pod) (int64, error) {
+	var output string
+	var err error
+
+	cmd := []string{"exec", pod.Name, "--namespace=" + namespace, "--", "/bin/sh", "-c", "df -Tkm | grep /mnt/volume1"}
+	output = e2ekubectl.RunKubectlOrDie(namespace, cmd...)
+	gomega.Expect(strings.Contains(output, ext4FSType)).NotTo(gomega.BeFalse())
+
+	arrMountOut := strings.Fields(string(output))
+	if len(arrMountOut) <= 0 {
+		return -1, fmt.Errorf("error when parsing output of `df -T`. output: %s", string(output))
+	}
+	var devicePath, strSize string
+	devicePath = arrMountOut[0]
+	if devicePath == "" {
+		return -1, fmt.Errorf("error when parsing output of `df -T` to find out devicePath of /mnt/volume1. output: %s",
+			string(output))
+	}
+	strSize = arrMountOut[2]
+	if strSize == "" {
+		return -1, fmt.Errorf("error when parsing output of `df -T` to find out size of /mnt/volume1: output: %s",
+			string(output))
+	}
+
+	intSizeInMb, err := strconv.ParseInt(strSize, 10, 64)
+	if err != nil {
+		return -1, fmt.Errorf("failed to parse size %s into int size", strSize)
+	}
+
+	return intSizeInMb, nil
+}
+
+// increaseSizeOfPvcAttachedToPod this method increases the PVC size, which is attached to POD
+func increaseSizeOfPvcAttachedToPodWithoutF(client clientset.Interface,
+	namespace string, pvclaim *v1.PersistentVolumeClaim, pod *v1.Pod) {
+	var originalSizeInMb int64
+	var err error
+	//Fetch original FileSystemSize
+	ginkgo.By("Verify filesystem size for mount point /mnt/volume1 before expansion")
+	originalSizeInMb, err = getFSSizeMbWithoutF(namespace, pod)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	//resize PVC
+	// Modify PVC spec to trigger volume expansion
+	ginkgo.By("Expanding current pvc")
+	currentPvcSize := pvclaim.Spec.Resources.Requests[v1.ResourceStorage]
+	newSize := currentPvcSize.DeepCopy()
+	newSize.Add(resource.MustParse("1Gi"))
+	framework.Logf("currentPvcSize %v, newSize %v", currentPvcSize, newSize)
+	pvclaim, err = expandPVCSize(pvclaim, newSize, client)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	gomega.Expect(pvclaim).NotTo(gomega.BeNil())
+
+	ginkgo.By("Waiting for file system resize to finish")
+	pvclaim, err = waitForFSResize(pvclaim, client)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	pvcConditions := pvclaim.Status.Conditions
+	expectEqual(len(pvcConditions), 0, "pvc should not have conditions")
+
+	var fsSize int64
+	ginkgo.By("Verify filesystem size for mount point /mnt/volume1")
+	fsSize, err = getFSSizeMbWithoutF(namespace, pod)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	framework.Logf("File system size after expansion : %v", fsSize)
+	// Filesystem size may be smaller than the size of the block volume
+	// so here we are checking if the new filesystem size is greater than
+	// the original volume size as the filesystem is formatted for the
+	// first time
+	gomega.Expect(fsSize).Should(gomega.BeNumerically(">", originalSizeInMb),
+		fmt.Sprintf("error updating filesystem size for %q. Resulting filesystem size is %d", pvclaim.Name, fsSize))
+	ginkgo.By("File system resize finished successfully")
+}
